qc/fastq_strand: relocated to 'qc.apps.fastq_strand' module.
