+++ auto_process_ngs/qc/plots.py
@@ -15,8 +15,8 @@ from bcftbx.utils import AttributeDictionary
 from .apps.fastqc import FastqcData
 from .apps.fastqc import FastqcSummary
 from .apps.fastq_screen import Fastqscreen
+from .apps.fastq_strand import Fastqstrand
 from .fastq_stats import FastqQualityStats
-from .fastq_strand import Fastqstrand
 
 # Data
 from .apps.fastqc import SEQUENCE_DEDUP_CUTOFFS
